+++ themes/ssp/default/includes/footer.php
@@ -1,6 +1,6 @@
 
 <?php
-$themeConfig = SimpleSAML_Configuration::getConfig('module_themevanilla.php');
+$themeConfig = SimpleSAML\Configuration::getConfig('module_themevanilla.php');
 $enable_cookies_banner = $themeConfig->getValue('enable_cookies_banner');
 
 if(!empty($this->data['htmlinject']['htmlContentPost'])) {
@@ -53,13 +53,13 @@ if ($includeLanguageBar) {
       'no' => 'Bokmål', // Norwegian Bokmål
       'nn' => 'Nynorsk', // Norwegian Nynorsk
       'se' => 'Sámegiella', // Northern Sami
-      'sam' => 'Åarjelh-saemien giele', // Southern Sami
       'da' => 'Dansk', // Danish
       'en' => 'English',
       'de' => 'Deutsch', // German
       'sv' => 'Svenska', // Swedish
       'fi' => 'Suomeksi', // Finnish
       'es' => 'Español', // Spanish
+      'ca' => 'Català', // Catalan
       'fr' => 'Français', // French
       'it' => 'Italiano', // Italian
       'nl' => 'Nederlands', // Dutch
@@ -80,14 +80,15 @@ if ($includeLanguageBar) {
       'zh' => '简体中文', // Chinese (simplified)
       'zh-tw' => '繁體中文', // Chinese (traditional)
       'ar' => 'العربية', // Arabic
-      'fa' => 'پارسی', // Persian
-      'ur' => 'اردو', // Urdu
       'he' => 'עִבְרִית', // Hebrew
       'id' => 'Bahasa Indonesia', // Indonesian
       'sr' => 'Srpski', // Serbian
       'lv' => 'Latviešu', // Latvian
       'ro' => 'Românește', // Romanian
       'eu' => 'Euskara', // Basque
+      'af' => 'Afrikaans', // Afrikaans
+      'zu' => 'IsiZulu', // Zulu
+      'xh' => 'isiXhosa', // Xhosa
     );
 
     $textarray = array();
@@ -96,7 +97,7 @@ if ($includeLanguageBar) {
       if ($current) {
         $lang_current = $langnames[$lang];
       } else {
-        $textarray[] = '<li class="ssp-dropdown__two_cols--item"><a href="' . htmlspecialchars(\SimpleSAML\Utils\HTTP::addURLParameters(\SimpleSAML\Utils\HTTP::getSelfURL(), array($this->languageParameterName => $lang))) . '">' .
+        $textarray[] = '<li class="ssp-dropdown__two_cols--item"><a href="' . htmlspecialchars(\SimpleSAML\Utils\HTTP::addURLParameters(\SimpleSAML\Utils\HTTP::getSelfURL(), array($this->getTranslator()->getLanguage()->getLanguageParameterName() => $lang))) . '">' .
           $langnames[$lang] . '</a></li>';
       }
     }
@@ -113,7 +114,7 @@ if ($includeLanguageBar) {
 ?>
     <div class="copy col-sm-6 ssp-footer__item">
       <a href="https://grnet.gr/">
-        <img class="ssp-footer__item__logo" src="<?php echo SimpleSAML_Module::getModuleURL('themevanilla/resources/images/grnet_logo_en.svg'); ?>" alt="GRNET" />
+        <img class="ssp-footer__item__logo" src="<?php echo SimpleSAML\Module::getModuleURL('themevanilla/resources/images/grnet_logo_en.svg'); ?>" alt="GRNET" />
       </a>
       <div class="ssp-footer__item__copyright">
         Copyright &copy;<?php echo (strpos($this->t('{themevanilla:discopower:copyright_year_start}'), 'not translated') === FALSE ? $this->t('{themevanilla:discopower:copyright_year_start}') . '-' : ''); echo date("Y"); ?>
@@ -127,19 +128,19 @@ if ($includeLanguageBar) {
   </div> <!-- /container-fluid -->
 </footer>
   <script type="text/javascript"
-          src="<?php echo htmlspecialchars(SimpleSAML_Module::getModuleURL('themevanilla/resources/js/cookie.js')); ?>">
+          src="<?php echo htmlspecialchars(SimpleSAML\Module::getModuleURL('themevanilla/resources/js/cookie.js')); ?>">
   </script>
   <script type="text/javascript"
-          src="<?php echo htmlspecialchars(SimpleSAML_Module::getModuleURL('themevanilla/resources/js/dropdown.js')); ?>">
+          src="<?php echo htmlspecialchars(SimpleSAML\Module::getModuleURL('themevanilla/resources/js/dropdown.js')); ?>">
   </script>
   <script type="text/javascript"
-          src="<?php echo htmlspecialchars(SimpleSAML_Module::getModuleURL('themevanilla/resources/js/modal.js')); ?>">
+          src="<?php echo htmlspecialchars(SimpleSAML\Module::getModuleURL('themevanilla/resources/js/modal.js')); ?>">
   </script>
   <script type="text/javascript"
-          src="<?php echo htmlspecialchars(SimpleSAML_Module::getModuleURL('themevanilla/resources/js/tooltip.js')); ?>">
+          src="<?php echo htmlspecialchars(SimpleSAML\Module::getModuleURL('themevanilla/resources/js/tooltip.js')); ?>">
   </script>
   <script type="text/javascript"
-          src="<?php echo htmlspecialchars(SimpleSAML_Module::getModuleURL('themevanilla/resources/js/theme.js')); ?>">
+          src="<?php echo htmlspecialchars(SimpleSAML\Module::getModuleURL('themevanilla/resources/js/theme.js')); ?>">
   </script>
 
 </body>