+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Add languages (Xhosa, Zulu, Afrikaans).
+
+### Changed
+- Rename disco-tpl.php->disco.tpl.php as needed by simplesamlphp>1.14.
+- Update translation functionality as needed by simplesamlphp>1.14.
+
 ## [v2.3.1](https://github.com/rciam/simplesamlphp-module-themevanilla/compare/v2.3.0...v2.3.1) - 2019-09-12
 
 ### Fixed